Fix types and use getEth2ApiClient function
